+++ build.extra.cjs
@@ -1,0 +1,19 @@
+const pkg = require('./package.json');
+
+module.exports = {
+  ...pkg.build,
+  extraResources: [
+    {
+      from: "plugins",
+      to: "plugins",
+      filter: [
+        "**/*",
+        "!**/node_modules",
+        "!**/.env",
+        "!**/scripts",
+        "!**/package.json",
+        "!**/package-lock.json"
+      ]
+    }
+  ]
+};